this was a bad idea
