+++ backend/scrapers/classes/main.js
@@ -17,7 +17,6 @@ import markMissingPrereqs from './processors/markMissingPrereqs';
 import termStartEndDate from './processors/termStartEndDate';
 import simplifyProfList from './processors/simplifyProfList';
 import addPreRequisiteFor from './processors/addPreRequisiteFor';
-import attachSectionsToClasses from './processors/attachSectionsToClasses';
 
 // Parsers
 import bannerv9Parser from './parsersxe/bannerv9Parser';
@@ -162,7 +161,6 @@ class Main {
     // Add new processors here.
     simplifyProfList.go(dump);
     addPreRequisiteFor.go(dump);
-    attachSectionsToClasses.go(dump);
 
     return dump;
   }
